issue #48: full separation of the two sides of the membrane via an environment registry
