+++ package.json
@@ -8,7 +8,8 @@
     "build": "ttsc && rollup -c",
     "start": "ttsc -w & http-server",
     "test": "jest --runInBand --no-cache --config ./jest.config.js",
-    "publish": "npm publish --access public"
+    "publish": "npm publish --access public",
+    "examples": "rollup -c"
   },
   "files": [
     "lib/",
